Fix 32-bit build failure due to `statFlagMap` using int keys
Change `statFlagMap` from `map[int]string` to `map[uintptr]string` and remove
int/uintptr casts in `statFlagNames()`, matching `mountFlagMap`. On 32-bit
architectures, `unix.MS_NOUSER` (0x80000000) overflows signed int.

Regression from PR #6129.

+++ chroot/run_linux.go
@@ -75,7 +75,7 @@ var (
 		unix.MS_SYNCHRONOUS:  "MS_SYNCHRONOUS",
 		unix.MS_UNBINDABLE:   "MS_UNBINDABLE",
 	}
-	statFlagMap = map[int]string{
+	statFlagMap = map[uintptr]string{
 		unix.ST_MANDLOCK:    "ST_MANDLOCK",
 		unix.ST_NOATIME:     "ST_NOATIME",
 		unix.ST_NODEV:       "ST_NODEV",
@@ -107,9 +107,9 @@ func statFlagNames(flags uintptr) []string {
 	var names []string
 	flags = flags & ^uintptr(0x20) // mask off ST_VALID
 	for flag, name := range statFlagMap {
-		if int(flags)&flag == flag {
+		if flags&flag == flag {
 			names = append(names, name)
-			flags = flags &^ (uintptr(flag))
+			flags = flags &^ flag
 		}
 	}
 	if flags != 0 { // got some unknown leftovers
